+++ src/Visitor/Stringify.php
@@ -65,6 +65,10 @@ use Typhoon\Type\TrueT;
 use Typhoon\Type\TruthyStringT;
 use Typhoon\Type\Type;
 use Typhoon\Type\UnionT;
+use Typhoon\Type\UntypedParameterT;
+use Typhoon\Type\UntypedPropertyT;
+use Typhoon\Type\UntypedReturnT;
+use Typhoon\Type\UntypedT;
 use Typhoon\Type\ValueOfT;
 use Typhoon\Type\Variance;
 use Typhoon\Type\Visitor;
@@ -633,6 +637,12 @@ abstract class Stringify implements Visitor
     public function mixedT(MixedT $type): string
     {
         return 'mixed';
+    }
+
+    #[\Override]
+    public function untypedT(UntypedT $type): mixed
+    {
+        return 'untyped-mixed';
     }
 
     /**